# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateExecutionRequest": {
        "properties": {
          "flowId": {
            "title": "Flowid",
            "type": "string"
          },
          "idempotencyKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotencykey"
          },
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "runner": {
            "$ref": "#/components/schemas/RunnerMode",
            "default": "local"
          }
        },
        "required": [
          "namespace",
          "flowId"
        ],
        "title": "CreateExecutionRequest",
        "type": "object"
      },
      "ExecutionDetail": {
        "properties": {
          "execution": {
            "$ref": "#/components/schemas/PersistedExecution"
          },
          "taskRuns": {
            "items": {
              "$ref": "#/components/schemas/PersistedTaskRun"
            },
            "title": "Taskruns",
            "type": "array"
          }
        },
        "required": [
          "execution",
          "taskRuns"
        ],
        "title": "ExecutionDetail",
        "type": "object"
      },
      "ExecutionEvent": {
        "properties": {
          "actor_id": {
            "default": "system",
            "title": "Actor Id",
            "type": "string"
          },
          "causation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causation Id"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/ExecutionEventType"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "event_type"
        ],
        "title": "ExecutionEvent",
        "type": "object"
      },
      "ExecutionEventType": {
        "enum": [
          "ExecutionCreated",
          "ExecutionQueued",
          "ExecutionStarted",
          "ExecutionPaused",
          "ExecutionResumed",
          "ExecutionCancelRequested",
          "ExecutionCancelled",
          "ExecutionSucceeded",
          "ExecutionFailed",
          "ExecutionWarned",
          "ExecutionRestartRequested"
        ],
        "title": "ExecutionEventType",
        "type": "string"
      },
...
      "ExecutionState": {
        "enum": [
          "CREATED",
          "QUEUED",
          "RUNNING",
          "PAUSED",
          "CANCELLING",
          "CANCELLED",
          "SUCCESS",
          "FAILED",
          "WARNING",
          "RESTARTING"
        ],
        "title": "ExecutionState",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PersistedExecution": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "epoch": {
            "minimum": 1.0,
            "title": "Epoch",
            "type": "integer"
          },
          "execution_id": {
            "format": "uuid",
            "title": "Execution Id",
            "type": "string"
          },
          "flow_id": {
            "title": "Flow Id",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "execution_id",
          "tenant_id",
          "state",
          "epoch",
          "version",
          "namespace",
          "flow_id",
          "created_at",
          "updated_at"
        ],
        "title": "PersistedExecution",
        "type": "object"
      },
      "PersistedFlow": {
        "properties": {
          "flow_id": {
            "title": "Flow Id",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "semantic_hash": {
            "title": "Semantic Hash",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "flow_id",
          "revision",
          "semantic_hash"
        ],
        "title": "PersistedFlow",
        "type": "object"
      },
      "PersistedTaskRun": {
        "properties": {
          "current_attempt": {
            "minimum": 0.0,
            "title": "Current Attempt",
            "type": "integer"
          },
          "execution_id": {
            "format": "uuid",
            "title": "Execution Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "retry_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry At"
          },
          "state": {
            "$ref": "#/components/schemas/TaskRunState"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_run_id": {
            "format": "uuid",
            "title": "Task Run Id",
            "type": "string"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_run_id",
          "execution_id",
          "task_id",
          "state",
          "current_attempt",
          "version"
        ],
        "title": "PersistedTaskRun",
        "type": "object"
      },
      "ReduceExecutionRequest": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEvent"
            },
            "minItems": 1,
            "title": "Events",
            "type": "array"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ExecutionSnapshot"
          }
        },
        "required": [
          "snapshot",
          "events"
        ],
        "title": "ReduceExecutionRequest",
        "type": "object"
      },
      "ReduceExecutionResponse": {
        "properties": {
          "duplicate_events_ignored": {
            "default": 0,
            "title": "Duplicate Events Ignored",
            "type": "integer"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ExecutionSnapshot"
          }
        },
        "required": [
          "snapshot"
        ],
        "title": "ReduceExecutionResponse",
        "type": "object"
      },
      "RunnerMode": {
        "enum": [
          "local",
          "kubernetes"
        ],
        "title": "RunnerMode",
        "type": "string"
      },
      "TaskLog": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "taskId",
          "attempt",
          "state"
        ],
        "title": "TaskLog",
        "type": "object"
      },
      "TaskRunState": {
        "enum": [
          "WAITING",
          "RUNNING",
          "RETRY_DELAY",
          "SUCCESS",
          "FAILED"
        ],
        "title": "TaskRunState",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Clean-room durable workflow MVP with validated flow management, execution control, webhook triggers and execution logs.",
    "title": "AMESH",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/executions": {
      "get": {
        "operationId": "list_executions_api_v1_executions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersistedExecution"
                  },
                  "title": "Response List Executions Api V1 Executions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Executions",
        "tags": [
          "executions"
        ]
      },
      "post": {
        "operationId": "create_execution_api_v1_executions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Execution",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/reduce": {
      "post": {
        "operationId": "reduce_execution_events_api_v1_executions_reduce_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReduceExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReduceExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reduce Execution Events",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/{execution_id}": {
      "get": {
        "operationId": "get_execution_api_v1_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Execution",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/logs": {
      "get": {
        "operationId": "get_execution_logs_api_v1_executions__execution_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskLog"
                  },
                  "title": "Response Get Execution Logs Api V1 Executions  Execution Id  Logs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Execution Logs",
        "tags": [
          "executions"
        ]
      }
    },
    "/api/v1/flows": {
      "get": {
        "operationId": "list_flows_api_v1_flows_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersistedFlow"
                  },
                  "title": "Response List Flows Api V1 Flows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Flows",
        "tags": [
          "flows"
        ]
      },
      "put": {
        "operationId": "apply_flow_api_v1_flows_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersistedFlow"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Flow",
        "tags": [
          "flows"
        ]
      }
    },
...
    "/api/v1/webhooks/{namespace}/{flow_id}/{trigger_id}": {
      "post": {
        "operationId": "trigger_webhook_api_v1_webhooks__namespace___flow_id___trigger_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "flow_id",
            "required": true,
            "schema": {
              "title": "Flow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "trigger_id",
            "required": true,
            "schema": {
              "title": "Trigger Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "runner",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunnerMode",
              "default": "local"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Webhook",
        "tags": [
          "triggers"
        ]
      }
    },
...
  }
}